Flutter Project - Before serving to Firebase
This is my commit prior to serving to firebase. This code works locally on my on system. It does not work well with full screen desktop view. If there is a black and yellow striped bar across the bottom, the navbar will not work. shrink window if that is the case and the bar works great.